Adding custom make_my_ray function
The make_my_ray function is the backbone of the circumgalactic dictionary. It has the same functionality as the make_simple_ray function, but also records metadata, particle data (and particle histories), black hole data, and ray geometry data. It records this data all in a single HDF5 file which can be loaded with (altered) yt