Added temporary fix to generating rays with observables
Currently a bug where some info is improperly recorded to the HDF5 file when rays are generated with observables=True. For now, we just do not attempt to record this data. Hopefully we do not skip any important lines, but if we do, the post-processing method seems to work properly